// src/api/generated/model/archive-package-vo.ts
// tslint:disable
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ArchivePackageVO
 */
export interface ArchivePackageVO {
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    creationTime?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ArchivePackageVO
     */
    deletable?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    projectId?: string;
    /**
     * 
     * @type {number}
     * @memberof ArchivePackageVO
     */
    tipVersion?: number;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    updateTime?: string;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof ArchivePackageVO
     */
    userName?: string;
}